Fix conclusion checks
Previously if a conclusion argument is an application expression, the
term being applied is checked only for whether it is a bound variable
(and if it is, an error is raised because bound variables can never be
functions). Now we also check if it is a valid bound var or extern
variable.